API

ParlinfoSpeechScraper

The ParlinfoSpeechScraper processes the xml files iteratively down the node tree. It reads the xml file from top down and processes one node at a time until it exhausts all content in the xml file.

ParlinfoSpeechScraper.get_args — Method

get_args()

Parse command-line arguments

This function sets up and parses command-line arguments using the ArgParse package. It defines the arguments that can be passed to the script, including options for verbosity and input file path.

ParlinfoSpeechScraper.main — Method

main(toml_path, verbose)

Main function to initialize and execute the Parlinfo Speech Scraper

This function initializes the input configuration from a TOML file specified by toml_path with optional verbosity controlled by verbose. It then runs the ParlinfoSpeechScraper using the configuration.

Inputs:

toml_path: Path to the TOML configuration file.

verbose: Boolean flag indicating if logging verbosity should be increased.

ParlinfoSpeechScraper.main — Method

main()

Main function to process command-line arguments and run the program

This function serves as the entry point of the script. It processes command-line arguments, retrieves the necessary parameters, and calls the main processing function with those parameters.

Inputs: None

Arguments:

args: A dictionary containing the parsed command-line arguments.

args["input"]: Path to the input TOML file.

args["verbose"]: Boolean flag indicating if logging verbosity should be increased.

RunModule

ParlinfoSpeechScraper.RunModule.get_date — Method

get_date(fn)

Get the date from the xml file

Inputs:

fn: the file directory for the xml file

ParlinfoSpeechScraper.RunModule.recurse — Function

recurse(soup, date, PhaseType, xmlnode, io, index=1, depth=0, maxdepth=0, node_tree=Vector{Node}())

Recursively process XML nodes and write data to output

This function recursively processes XML nodes, extracts relevant data, and writes it to an output stream. It handles node types, depth limitations, and maintains a tree of processed nodes.

Inputs:

soup: The parsed XML document.

date: The date associated with the XML document.

PhaseType: The phase type determined from the date.

xml_node: The current XML node being processed.

io: The output stream where data is written.

index (optional): The index of the current node (default is 1).

depth (optional): The current depth of recursion (default is 0).

max_depth (optional): The maximum depth for recursion (default is 0, meaning no limit).

node_tree (optional): A vector maintaining the tree of nodes (default is an empty vector).

ParlinfoSpeechScraper.RunModule.run_ParlinfoSpeechScraper — Method

run_ParlinfoSpeechScraper(toml::Dict{String, Any})

This function processes XML files for parliamentary speeches according to the configuration specified in the provided TOML dictionary. It reads XML file paths, processes each XML file, and outputs the results to a specified directory.

Inputs:

toml: A dictionary containing configuration options for the scraper.

ParlinfoSpeechScraper.RunModule.run_xml — Method

run_xml(fn, output_path, csv_exist, edit_opt)

Process and save XML data to CSV

This function processes an XML file, extracts relevant data, and saves it to a CSV file. If specified, it also edits the CSV file after creation.

Inputs:

fn: The file path for the XML file.

output_path: The directory where the processed CSV file will be saved.

csv_exist: Boolean flag indicating if a CSV file already exists.

edit_opt: Boolean flag indicating if the CSV file should be edited after creation.

XMLModule

ParlinfoSpeechScraper.RunModule.EditModule.edit_csv — Method

edit_csv(fn, ::Type{<:AbstractPhase})

Two steps are implemented here. The first step does mostly regular expression processing. The second step collapses some of the N/As to the previous talker.

ParlinfoSpeechScraper.RunModule.EditModule.edit_set_up — Method

editsetup(headers)

Sets up a dictionary mapping each element in headers to its corresponding index.

Inputs:

headers: An array of header names or keys.

Returns:

A dictionary where each header in headers is mapped to its index position.

ParlinfoSpeechScraper.RunModule.EditModule.find_all_child_speeches — Method

function findallchildspeeches(rowno,rows,headertonum,is_written)

Find all the speeches that belong to a single talker