Nodes

NodeModule

ParlinfoSpeechScraper.RunModule.NodeModule.detect_phase — Method

detect_phase(date)

Inputs:

date: A floating-point number representing the date to detect the phase for.

Returns:

The phase associated with the provided date, or AbstractPhase if no specific phase is found.

ParlinfoSpeechScraper.RunModule.NodeModule.get_all_subtypes — Function

get_all_subtypes(type, st=[])

Get every subtype of the provided type parameter.

Inputs:

type: the provided type

ParlinfoSpeechScraper.RunModule.NodeModule.reverse_find_first_node_name — Method

reversefindfirstnodename(node_tree, names)

Inputs:

node_tree: A vector representing a tree of nodes to search in reverse order.

names: A collection of node names to search for.

Returns:

The first node from node_tree in reverse order whose name is in the names collection, or nothing if no such node is found.

ParlinfoSpeechScraper.RunModule.NodeModule.reverse_find_first_node_not_name — Method

reversfindfirstnodenotname(nodetree, names)

Inputs:

node_tree: A vector representing a tree of nodes to search in reverse order.

names: A collection of node names. The function searches for the first node from node_tree in reverse order whose name is not in this collection.

Returns:

The first node from node_tree in reverse order whose name is not in the names collection, or nothing if no such node is found.

ParlinfoSpeechScraper.RunModule.NodeModule.detect_node_type — Method

detectnodetype(node, node_tree, date, soup, PhaseType)

Inputs:

node: The XML node to determine the type for.

node_tree: A vector representing a tree of nodes

date: The date associated with the node.

soup: The root node.

PhaseType: The phase type determined from the date of XML.

Returns:

The detected node type (NodeType).

ParlinfoSpeechScraper.RunModule.NodeModule.parse_node — Method

parsenode(node::Node, nodetree, io)

Inputs:

node: xml node of Node struct

node_tree: A vector representing a tree of nodes for context.

io: The output stream where processed data is written.

ParlinfoSpeechScraper.RunModule.NodeModule.process_node — Method

processnode(node::Node, nodetree)

This function serves as a placeholder or default behavior.

Inputs:

node: xml node of Node struct

node_tree: A vector representing a tree of nodes for context.

Notes:

This function is typically invoked when no specific processing behavior is defined for the NodeType associated with node.

Users may customize or define specific behaviors for different NodeType instances within their implementation.

ParlinfoSpeechScraper.RunModule.NodeModule.is_nodetype — Method

isnodetype(node, nodetree, nodetype::Type{<:AbstractNode}, phase::Type{<:AbstractPhase}, soup, args...; kwargs...)

Inputs:

node: The XML node to evaluate.

node_tree: A vector representing a tree of nodes for context.

nodetype: A subtype of AbstractNode representing the type of node to check against.

phase: A subtype of AbstractPhase representing the phase associated with the node.

soup: The root node.

args...: Optional additional arguments.

kwargs...: Optional keyword arguments.

Returns:

true if the name of node is found in the allowed names associated with nodetype for the given phase, false otherwise.

Notes: More specific detection method for nodes is defined in nodes/

ParlinfoSpeechScraper.RunModule.NodeModule.get_xpaths — Method

get_xpaths(::Type{<:N}) where {N <: AbstractNode}

This function serves as a placeholder and returns an empty array, which is meant to provide the nodenames allowed for each type. For example, speech nodes can have names "speech" or "continue".

ParlinfoSpeechScraper.RunModule.NodeModule.find_section_title — Method

findsectiontitle(node, nodetree, soup, sectiontype)

Inputs:

node: The XML node from which to extract the section title.

node_tree: A vector representing a tree of nodes for context.

soup: The root node of the XML document.

section_type: The nodenames of the types of section nodes where the title is wanted. For example, "speech".

Returns:

The title of the specified section_type found within the XML document, or "N/A" if not found.

ParlinfoSpeechScraper.RunModule.NodeModule.find_chamber — Method

findchamber(node, nodetree)

Identifies the type of chamber associated with the given XML node.

Inputs:

node: The XML node

node_tree: A vector representing a tree of nodes for context.

Returns:

An integer indicating the chamber type:

2 for a federal chamber (FedChamberNode).

1 for a chamber (ChamberNode).

0 if no chamber node is found.

ParlinfoSpeechScraper.RunModule.NodeModule.define_flags — Method

defineflags(node::Node{<:AbstractNode{<:AbstractPhase}}, nodetree)

Generates flags based on the characteristics of the given node.

Inputs:

node: A Node struct.

node_tree: A vector representing a tree of nodes for context.

Returns:

An array of flags indicating characteristics of node

ParlinfoSpeechScraper.RunModule.NodeModule.construct_row — Method

constructrow(node, nodetree, flags, talker_contents, content)

Inputs:

node: A Node struct.

node_tree: A vector representing a tree of nodes for context.

flags: An array of flags indicating characteristics of node.

talker_contents: Contents related to speakers or talkers associated with node.

content: Raw content associated with node.

Returns:

An array (row) representing the data to be written to a CSV file. The row includes:

Flags indicating characteristics of node.

Talker contents related to speakers or talkers.

Cleaned text content derived from content.

Section title for SubdebateNode, if found.

Section title for DebateNode, if found.

XPath path of node within the XML document.

2012-2024 (default)

Here we show how the nodes are defined for different years.

PNode

xpaths: "p"

ParlinfoSpeechScraper.RunModule.NodeModule.get_xpaths — Method

get_xpaths(::Type{<:PNode})

Find what the p_node is called in xml for default

ParlinfoSpeechScraper.RunModule.NodeModule.process_node — Method

processnode(node::Node{<:PNode},nodetree)

If no particular phase is specified, this default version of process node is run.

ParlinfoSpeechScraper.RunModule.NodeModule.is_first_node_type — Method

isfirstnodetype(node::Node{<:PNode},parentnode,allowednames,nodetree)

This default function detects if the pnode detected is the first pnode under a parent node. The reason for this is that if the p_node is not the first node, there are alternative checks to see if the talker has changed since the first one.

ParlinfoSpeechScraper.RunModule.NodeModule.find_talker_in_p — Method

findtalkerinp(pnode)

If the pnode is not the first pnode, we check if there is a talker inside the p_node.

ParlinfoSpeechScraper.RunModule.NodeModule.p_with_a_as_parent — Method

pwithaasparent(p_node)

If no talker is found for the first node, we look for a "a" as parentnode.

ParlinfoSpeechScraper.RunModule.NodeModule.get_sections — Method

get_sections(::Type{<:PNode})

In which sections are the p_node wanted as default.

ParlinfoSpeechScraper.RunModule.NodeModule.is_nodetype — Method

isnodetype(node, nodetree, nodetype::Type{<:PNode},phase::Type{<:AbstractPhase},soup, args...; kwargs...)

This function detects whether the given xml node is PNode. This function takes the phase into account aswell.

ParlinfoSpeechScraper.RunModule.NodeModule.parse_node — Method

parsenode(node::Node{<:PNode},nodetree,io)

The default function to take a p_node and write to the csv

ParlinfoSpeechScraper.RunModule.NodeModule.get_talker_from_parent — Method

gettalkerfromparent(::Type{<:PNode},parentnode)

If the pnode is the first pnode, we search in the parent to find the talker.

InterTalkNode

xpaths: "talk.start"

ParlinfoSpeechScraper.RunModule.NodeModule.get_xpaths — Method

get_xpaths(::Type{<:InterTalkNode})

The default setting for what nodenames are allowed for intertalknode.

ParlinfoSpeechScraper.RunModule.NodeModule.process_node — Method

processnode(node::Node{<:InterTalkNode},nodetree)

The default setting for processing intertalknodes.

ParlinfoSpeechScraper.RunModule.NodeModule.get_sections — Method

get_sections(::Type{<:InterTalkNode})

The default setting for the sections where intertalknodes are processed

ParlinfoSpeechScraper.RunModule.NodeModule.is_nodetype — Method

isnodetype(node, nodetree, nodetype::Type{<:InterTalkNode},phase::Type{<:AbstractPhase},soup, args...; kwargs...)

This function checks if the given xml node is of nodetype InterTalkNode

ParlinfoSpeechScraper.RunModule.NodeModule.get_talker_from_parent — Method

gettalkerfromparent(::Type{InterTalkNode},parentnode)

Get the talker information from the parentnode.

Placeholder nodes

AnswerNode

xpaths: "answer"

BusinessNode

xpaths: "business.start"

ChamberNode

xpaths: "chamber.xscript"

FedChamberNode

xpaths: "fedchamb.xscript", "maincomm.xscript"

DebateNode

xpaths: "debate"

DivisionNode

xpaths: "division"

InterjectionNode

xpaths: "interjection", "continue"

QuestionNode

xpaths: "question"

SpeechNode

xpaths: "speech", "question", "answer"

SubdebateNode

xpaths: "subdebate.1"

before April 2011

Phase information:

Phase2011

ParlinfoSpeechScraper.RunModule.NodeModule.define_flags — Method

defineflags(node::Node{<:AbstractNode{Phase2011}},parentnode,node_tree)

Added petition and quote nodes on top of the previous flags

PNode

xpaths: "para"

ParlinfoSpeechScraper.RunModule.NodeModule.get_xpaths — Method

get_xpaths(::Type{PNode{Phase2011}})

"para" as accepted xpaths for PNode.

ParlinfoSpeechScraper.RunModule.NodeModule.get_sections — Method

get_sections(::Type{PNode{Phase2011}})

Allowed sections for PNodes.

ParlinfoSpeechScraper.RunModule.NodeModule.is_first_node_type — Method

isfirstnodetype(node::Node{PNode{Phase2011}},parentnode,allowednames,nodetree)

A different method to detect if the pnode is the first pnode for this phase

Placeholder nodes

MotionnospeechNode

xpaths: "motionnospeech"

PetitionNode

xpaths: "petition"

QuoteNode_

xpaths: "quote"