Function references

ParlinfoSpeechScraper.get_args — Method

get_args()

Parse command-line arguments

This function sets up and parses command-line arguments using the ArgParse package. It defines the arguments that can be passed to the script, including options for verbosity and input file path.

ParlinfoSpeechScraper.main — Method

main(toml_path, verbose)

Main function to initialize and execute the Parlinfo Speech Scraper

This function initializes the input configuration from a TOML file specified by toml_path with optional verbosity controlled by verbose. It then runs the ParlinfoSpeechScraper using the configuration.

Inputs:

toml_path: Path to the TOML configuration file.

verbose: Boolean flag indicating if logging verbosity should be increased.

ParlinfoSpeechScraper.main — Method

main()

Main function to process command-line arguments and run the program

This function serves as the entry point of the script. It processes command-line arguments, retrieves the necessary parameters, and calls the main processing function with those parameters.

RunModule

ParlinfoSpeechScraper.RunModule.get_date — Method

get_date(fn)

Get the date from the xml file

Inputs:

fn: the file directory for the xml file

ParlinfoSpeechScraper.RunModule.recurse — Function

recurse(soup, date, PhaseType, xmlnode, io, index=1, depth=0, maxdepth=0, node_tree=Vector{Node}())

Recursively process XML nodes and write data to output

This function recursively processes XML nodes, extracts relevant data, and writes it to an output stream. It handles node types, depth limitations, and maintains a tree of processed nodes.

Inputs:

soup: The parsed XML document.

date: The date associated with the XML document.

PhaseType: The phase type determined from the date.

xml_node: The current XML node being processed.

io: The output stream where data is written.

index (optional): The index of the current node (default is 1).

depth (optional): The current depth of recursion (default is 0).

max_depth (optional): The maximum depth for recursion (default is 0, meaning no limit).

node_tree (optional): A vector maintaining the tree of nodes (default is an empty vector).

ParlinfoSpeechScraper.RunModule.run_ParlinfoSpeechScraper — Method

run_ParlinfoSpeechScraper(toml::Dict{String, Any})

This function processes XML files for parliamentary speeches according to the configuration specified in the provided TOML dictionary. It reads XML file paths, processes each XML file, and outputs the results to a specified directory.

Inputs:

toml: A dictionary containing configuration options for the scraper.

ParlinfoSpeechScraper.RunModule.run_xml — Method

run_xml(fn, output_path, csv_exist, edit_funcs)

Process and save XML data to CSV

This function processes an XML file, extracts relevant data, and saves it to a CSV file. If specified, it also edits the CSV file after creation.

Inputs:

fn: The file path for the XML file.

output_path: The directory where the processed CSV file will be saved.

csv_exist: Boolean flag indicating if a CSV file already exists.

edit_funcs: list of edit functions

XMLModule

ParlinfoSpeechScraper.RunModule.EditModule.edit_main — Method

func_list: a list of function names as string editor: a struct with two parameters

ParlinfoSpeechScraper.RunModule.EditModule.edit_set_up — Method

editsetup(headers)

Sets up a dictionary mapping each element in headers to its corresponding index.

ParlinfoSpeechScraper.RunModule.EditModule.find_all_child_speeches — Method

function findallchildspeeches(rowno,rows,headertonum,is_written)

Find all the speeches that belong to a single talker

ParlinfoSpeechScraper.RunModule.EditModule.find_all_flags — Method

get all flags except chamber flag

ParlinfoSpeechScraper.RunModule.EditModule.free_node_op — Method

if it is free flowing, check if it is same debate and add it to the previous one

NodeModule

ParlinfoSpeechScraper.RunModule.NodeModule.define_flags — Method

defineflags(node::Node{<:AbstractNode{<:AbstractPhase}}, nodetree)

Generates flags based on the characteristics of the given node.

Inputs:

node: A Node struct.

node_tree: A vector representing a tree of nodes for context.

Returns:

An array of flags indicating characteristics of node

ParlinfoSpeechScraper.RunModule.NodeModule.define_flags — Method

defineflags(node::Node{<:AbstractNode{Phase2011}},parentnode,node_tree)

Added petition and quote nodes on top of the previous flags

ParlinfoSpeechScraper.RunModule.NodeModule.detect_node_type — Method

detectnodetype(node, node_tree, date, soup, PhaseType)

default setting:

Inputs:

node: The XML node to determine the type for.

node_tree: A vector representing a tree of nodes

soup: The root node.

PhaseType: The phase type determined from the date of XML.

Returns:

The detected node type (NodeType).

ParlinfoSpeechScraper.RunModule.NodeModule.detect_phase — Method

detectphase(date, whichhouse)

Inputs:

date: A floating-point number representing the date to detect the phase for.

house: A string that indicates either senate or house

Returns:

The phase associated with the provided date, or AbstractPhase if no specific phase is found.

ParlinfoSpeechScraper.RunModule.NodeModule.find_chamber — Method

findchamber(node, nodetree)

Identifies the type of chamber associated with the given XML node.

Inputs:

node: The XML node

node_tree: A vector representing a tree of nodes for context.

Returns:

An integer indicating the chamber type:

2 for a federal chamber (FedChamberNode).

1 for a chamber (ChamberNode).

0 if no chamber node is found.

ParlinfoSpeechScraper.RunModule.NodeModule.find_section_title — Method

findsectiontitle(nodetree, soup, sectiontype)

Inputs:

node: The XML node from which to extract the section title.

node_tree: A vector representing a tree of nodes for context.

soup: The root node of the XML document.

section_type: The nodenames of the types of section nodes where the title is wanted. For example, "speech".

Returns:

The title of the specified section_type found within the XML document, or "N/A" if not found.

ParlinfoSpeechScraper.RunModule.NodeModule.find_talker_in_p — Method

findtalkerinp(pnode)

If the pnode is not the first pnode, we check if there is a talker inside the p_node.

ParlinfoSpeechScraper.RunModule.NodeModule.get_all_subtypes — Function

get_all_subtypes(type, st=[])

Get every subtype of the provided type parameter.

Inputs:

type: the provided type

ParlinfoSpeechScraper.RunModule.NodeModule.get_sections — Method

get_sections(::Type{<:PNode})

In which sections are the p_node wanted as default.

ParlinfoSpeechScraper.RunModule.NodeModule.get_sections — Method

get_sections(::Type{PNode{Phase2011}})

Allowed sections for PNodes.

ParlinfoSpeechScraper.RunModule.NodeModule.get_sections — Method

get_sections(::Type{PNode{PhaseSGML}})

Allowed sections for PNodes.

ParlinfoSpeechScraper.RunModule.NodeModule.get_talker_from_parent — Method

gettalkerfromparent(node::Node,parentnode::Node)

It finds the speaker information from the parent node

ParlinfoSpeechScraper.RunModule.NodeModule.get_xpaths — Method

get_xpaths(::Type{<:InterTalkNode})

The default setting for what nodenames are allowed for intertalknode.

ParlinfoSpeechScraper.RunModule.NodeModule.get_xpaths — Method

get_xpaths(::Type{<:PNode})

Find what the p_node is called in xml for default

ParlinfoSpeechScraper.RunModule.NodeModule.get_xpaths — Method

get_xpaths(::Type{PNode{Phase2011}})

"para" as accepted xpaths for PNode.

ParlinfoSpeechScraper.RunModule.NodeModule.get_xpaths — Method

get_xpaths(::Type{PNode{PhaseSGML}})

"para" as accepted xpaths for PNode.

ParlinfoSpeechScraper.RunModule.NodeModule.get_xpaths — Method

get_xpaths(::Type{<:N}) where {N <: AbstractNode}

This function serves as a placeholder and returns an empty array, which is meant to provide the nodenames allowed for each type. For example, speech nodes can have names "speech" or "continue".

ParlinfoSpeechScraper.RunModule.NodeModule.is_first_node_type — Method

isfirstnodetype(node::Node{<:PNode},parentnode,allowednames,nodetree)

This default function detects if the pnode detected is the first pnode under a parent node. The reason for this is that if the p_node is not the first node, there are alternative checks to see if the talker has changed since the first one.

ParlinfoSpeechScraper.RunModule.NodeModule.is_first_node_type — Method

isfirstnodetype(node::Node{PNode{Phase2011}},parentnode,allowednames,nodetree)

A different method to detect if the pnode is the first pnode for this phase

ParlinfoSpeechScraper.RunModule.NodeModule.is_nodetype — Method

isnodetype(node, nodetree, nodetype::Type{<:PNode},phase::Type{<:AbstractPhase},soup, args...; kwargs...)

This function detects whether the given xml node is PNode. This function takes the phase into account aswell.

ParlinfoSpeechScraper.RunModule.NodeModule.is_nodetype — Method

isnodetype(node, nodetree, nodetype::Type{<:AbstractNode}, phase::Type{<:AbstractPhase}, soup, args...; kwargs...)

default setting:

Inputs:

node: The XML node to evaluate.

node_tree: A vector representing a tree of nodes for context.

nodetype: A subtype of AbstractNode representing the type of node to check against.

phase: A subtype of AbstractPhase representing the phase associated with the node.

soup: The root node.

args...: Optional additional arguments.

kwargs...: Optional keyword arguments.

Returns:

true if the name of node is found in the allowed names associated with nodetype for the given phase, false otherwise.

Notes: More specific detection method for nodes is defined in nodes/

ParlinfoSpeechScraper.RunModule.NodeModule.p_with_a_as_parent — Method

pwithaasparent(p_node)

If no talker is found for the first node, we look for a "a" as parentnode.

ParlinfoSpeechScraper.RunModule.NodeModule.parse_node — Method

parsenode(node::Node, nodetree, io)

default setting:

Inputs:

node: xml node of Node struct

node_tree: A vector representing a tree of nodes for context.

io: The output stream where processed data is written.

ParlinfoSpeechScraper.RunModule.NodeModule.parse_node — Method

parsenode(node::Node{<:PNode},nodetree,io)

The default function to take a p_node and write to the csv

ParlinfoSpeechScraper.RunModule.NodeModule.process_node — Method

processnode(node::Node, nodetree)

default setting:

Inputs:

node: xml node of Node struct

node_tree: A vector representing a tree of nodes for context.

Notes:

This function is typically invoked when no specific processing behavior is defined for the NodeType associated with node.

Users may customize or define specific behaviors for different NodeType instances within their implementation.

ParlinfoSpeechScraper.RunModule.NodeModule.process_node — Method

processnode(node::Node{<:PNode},nodetree)

If no particular phase is specified, this default version of process node is run.

ParlinfoSpeechScraper.RunModule.NodeModule.reverse_find_first_node_name — Method

reversefindfirstnodename(node_tree, names)

Inputs:

node_tree: A vector representing a tree of nodes to search in reverse order.

names: A collection of node names to search for.

Returns:

The first node from node_tree in reverse order whose name is in the names collection, or nothing if no such node is found.

ParlinfoSpeechScraper.RunModule.NodeModule.reverse_find_first_node_not_name — Method

reversfindfirstnodenotname(nodetree, names)

Inputs:

node_tree: A vector representing a tree of nodes to search in reverse order.

names: A collection of node names. The function searches for the first node from node_tree in reverse order whose name is not in this collection.

Returns:

The first node from node_tree in reverse order whose name is not in the names collection, or nothing if no such node is found.

ParlinfoSpeechScraper.RunModule.NodeModule.write_test_xml — Method

Writes the test xmls for all edge cases: it cannot handle dummy parent as root itself – soup is getting passed and not document type