Nodes

2012-2024 (default)

Here we show how the nodes are defined for different years.

PNode

xpaths: "p"

ParlinfoSpeechScraper.RunModule.NodeModule.get_xpaths — Method

get_xpaths(::Type{<:PNode})

Find what the p_node is called in xml for default

ParlinfoSpeechScraper.RunModule.NodeModule.process_node — Method

processnode(node::Node{<:PNode},nodetree)

If no particular phase is specified, this default version of process node is run.

ParlinfoSpeechScraper.RunModule.NodeModule.is_first_node_type — Method

isfirstnodetype(node::Node{<:PNode},parentnode,allowednames,nodetree)

This default function detects if the pnode detected is the first pnode under a parent node. The reason for this is that if the p_node is not the first node, there are alternative checks to see if the talker has changed since the first one.

ParlinfoSpeechScraper.RunModule.NodeModule.find_talker_in_p — Method

findtalkerinp(pnode)

If the pnode is not the first pnode, we check if there is a talker inside the p_node.

ParlinfoSpeechScraper.RunModule.NodeModule.p_with_a_as_parent — Method

pwithaasparent(p_node)

If no talker is found for the first node, we look for a "a" as parentnode.

ParlinfoSpeechScraper.RunModule.NodeModule.get_sections — Method

get_sections(::Type{<:PNode})

In which sections are the p_node wanted as default.

ParlinfoSpeechScraper.RunModule.NodeModule.is_nodetype — Method

isnodetype(node, nodetree, nodetype::Type{<:PNode},phase::Type{<:AbstractPhase},soup, args...; kwargs...)

This function detects whether the given xml node is PNode. This function takes the phase into account aswell.

ParlinfoSpeechScraper.RunModule.NodeModule.parse_node — Method

parsenode(node::Node{<:PNode},nodetree,io)

The default function to take a p_node and write to the csv

ParlinfoSpeechScraper.RunModule.NodeModule.get_talker_from_parent — Method

gettalkerfromparent(::Type{<:PNode},parentnode)

If the pnode is the first pnode, we search in the parent to find the talker.

InterTalkNode

xpaths: "talk.start"

ParlinfoSpeechScraper.RunModule.NodeModule.get_xpaths — Method

get_xpaths(::Type{<:InterTalkNode})

The default setting for what nodenames are allowed for intertalknode.

ParlinfoSpeechScraper.RunModule.NodeModule.process_node — Method

processnode(node::Node{<:InterTalkNode},nodetree)

The default setting for processing intertalknodes.

ParlinfoSpeechScraper.RunModule.NodeModule.get_sections — Method

get_sections(::Type{<:InterTalkNode})

The default setting for the sections where intertalknodes are processed

ParlinfoSpeechScraper.RunModule.NodeModule.is_nodetype — Method

isnodetype(node, nodetree, nodetype::Type{<:InterTalkNode},phase::Type{<:AbstractPhase},soup, args...; kwargs...)

This function checks if the given xml node is of nodetype InterTalkNode

ParlinfoSpeechScraper.RunModule.NodeModule.get_talker_from_parent — Method

gettalkerfromparent(::Type{InterTalkNode},parentnode)

Get the talker information from the parentnode.

Placeholder nodes

AnswerNode

xpaths: "answer"

BusinessNode

xpaths: "business.start"

ChamberNode

xpaths: "chamber.xscript"

FedChamberNode

xpaths: "fedchamb.xscript", "maincomm.xscript"

DebateNode

xpaths: "debate"

DivisionNode

xpaths: "division"

InterjectionNode

xpaths: "interjection", "continue"

QuestionNode

xpaths: "question"

SpeechNode

xpaths: "speech", "question", "answer"

SubdebateNode

xpaths: "subdebate.1"

before April 2011

Phase information:

Phase2011

ParlinfoSpeechScraper.RunModule.NodeModule.define_flags — Method

defineflags(node::Node{<:AbstractNode{Phase2011}},parentnode,node_tree)

Added petition and quote nodes on top of the previous flags

PNode

xpaths: "para"

ParlinfoSpeechScraper.RunModule.NodeModule.get_xpaths — Method

get_xpaths(::Type{PNode{Phase2011}})

"para" as accepted xpaths for PNode.

ParlinfoSpeechScraper.RunModule.NodeModule.get_sections — Method

get_sections(::Type{PNode{Phase2011}})

Allowed sections for PNodes.

ParlinfoSpeechScraper.RunModule.NodeModule.is_first_node_type — Method

isfirstnodetype(node::Node{PNode{Phase2011}},parentnode,allowednames,nodetree)

A different method to detect if the pnode is the first pnode for this phase

Placeholder nodes

MotionnospeechNode

xpaths: "motionnospeech"

PetitionNode

xpaths: "petition"

QuoteNode_

xpaths: "quote"